Add Avatar and Card base UI components
All "Rx" components are base components and can be reused easily throughout the app.
Avatar - diplays an image of a specific size. Down the line these sizes should be props and generated.
Card - White background to stay consistent throughout the App. Basic display of cards within the app, content is whatever you want it to be