Add Worker Components
SwipeWorkers - This is the top level component that manages the data. It will be responsible for generating the workers and updating the users choice on whether they swiped left or right.

WorkerCard - This component displays the WorkerCard, and manages swipes. It will call the prop right when it is 3/4 across the screen or left when it less than 1/4 across the screen